andrade i start for att anvanda riktig gps koordinat

+++ kartkomponent/start2.py
@@ -15,6 +15,7 @@ class Start(object):
 		self.coord = None
 		self.gpsrun = True
 		self.stringcoord = None
+		self.tank_added = False
 		
 	def to_tuple(self, stringen):
 		tupeln = tuple(stringen.split())
@@ -60,14 +61,19 @@ class Start(object):
 		while(self.gpsrun == True):		
 			#self.stringcoord = self.osso_rpc.rpc_run("thor.tufftuff", "/thor/tufftuff", "thor.tufftuff", "updatecoord", (), wait_reply = True)
 			self.stringcoord = self.osso_rpc.rpc_run("thor.gps", "/thor/gps", "thor.gps", "updatecoord", (), wait_reply = True)
+			print self.stringcoord
+			time.sleep(5)
+
 			
 			if(self.stringcoord != 0):
+				print "ritar ut tanken nu"
 				self.coord = self.to_tuple(self.stringcoord)
+				if(self.tank_added == True):
+					self.map.delete_object("Tank")
 				self.map.add_object("Tank", data_storage.MapObject({"longitude":(self.coord[1]-0.0016),
 									"latitude":(self.coord[0]+0.00075)},
 									"ikoner/tank.png"))
-				time.sleep(5)
-				self.map.delete_object("Tank")
+				self.tank_added = True
 			
 	def startgui(self):
 		self.gui = guitest.Gui(self.map)
